requirements.txt
txtFlask==3.0.0
python-pptx==0.6.23
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0
